Move plugin loader instance creation (#71)
The `$pluginLoader` does not need to be added as a property on the application instance. This pull request simply moves it to the `run` method. 

Since this was available on the application instance as a public property (visibility not specified) we should regard this as a breaking change and merge it in version 7.0.